+++ backend/api/serializers.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from event.models import Event
+from event.models.models import Event
 
 class EventSerializer(serializers.ModelSerializer):
     class Meta:
